Update to ESLint 3.
Now that ESLint supports only Node versions > 0.12, we should
start to begin our transition to Node 4. Right now, this means
that we are dropping linting on 0.12 builds from Travis.
Furthermore, 0.12 is being dropped as a test environment from
Appveyor.

Module code for the 3.x releases will still continue to work
on 0.12 until the next major version.

Closes #216.
Closes #217.
Closes #218.
Closes #221.
Closes #222.
Closes #223.
Closes #224.