Don't adjust the semantics of namespace URLs in any way.
Namespace identifiers should not be transformed; using a URI
(such as "http://www.w3.org/1999/xhtml") to identify a namespace,
rather than a simple string (such as "xhtml"), reduces the
probability of different namespaces using duplicate identifiers.

https://en.wikipedia.org/wiki/XML_namespace

This patch ensures that the only namespace transform that is
performed is the conversion from a url(foo) function to a URL
string "foo".