Fix interoperability issue with modular-css.
The previous behaviour would use the raw selector string, which
is OK if previous PostCSS plugins update this when setting the
new value; but for those that don't, the result is as if the
plugin didn't run at all.

The new behaviour tests that the raw value is the same as the
selector's 'cleaned' value; if it is not, then that means that
another plugin has run on the selector string and we should not
use the raw value.

Reference: https://github.com/tivac/modular-css/issues/228