refactor(postcss-colormin): Improve caching.
Previously a cache was kept per property/value pair as well as
the transformed value. This meant that otherwise identical values
would result in a cache miss. Instead, we now cache at the value
level and at the colour transform level, which should improve
performance.

This patch is also experimenting with a more declarative style
using Ramda in order to do the caching, and some small quality
of life fixes.