Move postcss to dependencies
When postcss is not installed as a direct dependency in a project, yarn
may print a bunch of warnings regarding the missing peer dependency.
Move it to an actual dependency to avoid those warnings.

Behaviour with npm v7 is to automatically install peerDependencies so it
makes no practical difference whether a module is in dependencies vs.
peerDependencies.