fix(postcss-merge-longhand): skip processing if diff not exactly 1
Fix #1217

Before we skipped if the diff was larger than 1, this excluded the case where the diff is 0.
Actually if the declarations are identical this part of the code should probably never execute in
the first place, but the issue only triggers with very specific CSS combinations so I could not
find a different fix.