chore: remove postcss-devtools (#1229)
I don't think this postcss-devtools has any effect now.
According to its README, postcss-devtools is supposed to measure the running time
of individuals plugins. postcss-devtools has not been updated since PostCSS 5.
It was used in a single test but the test does not inspect the running time
and I can't see the running time in the test output either.