refactor(postcss-normalize-display-values): improve transformation lookup
Apply transform by using a map from the original value to the
transformed value. Stringify the original value array to use as a
key.

Refactor the test so it uses an independent mapping, otherwise
it will pass even if the mapping in the source code is wrong.

Remove cssnano-utils dependency.