feat!(postcss-normalize-url): inline third-party dep and remove options (#1480)
* chore: bump ECMAscript version

Support the normalize-url code.

* feat!(postcss-normalize-url): inline third-party dep and remove options

Inline a simplified version of normalize-url. Reasons:
- the two most recent major releases of normalize-url use ES modules,
  so cssnano cannot use them
- most options change the meaning of the URLs, so it is unlikely
  that turning them on during minification makes sense

THe remaining code removes redundant slashes and default ports,
so performs the same as the previous default configuration. It does not
sort parameters any more because we haven't yet found a method that preserves
the correct encoding in all cases.
If the user does not like these transforms they can turn the plugin
off completely.

* docs(postcss-normalize-url): update readme