fix(cssnano): handle cases where AtRules might not have children
Since the introduction of the @layer AtRule, valid AtRules might lack children.
PostCSS 8.35 has updated types to reflect that, so we now need to check whether
children exist before accessing them. In many cases it would not be necessary,
since we already process specific atRules and we know they are not @layer,
but the types are not fine-grained enough to relfect that.

Fix #1510