perf: load default preset in startup
The default preset is a widely relied on, especially by frameworks (e.g.
nuxt).

Currently, this means the hot path will have to `require` the preset and
_block until resolution finishes_.

If we load the default preset up front, it means the call to cssnano
will actually run much faster.

This does mean initial load is likely technically slower, but the gain
of the call performance seems to outweigh this.